Fixes filename and folder problem
Previously file were just getting saved in the root directory of
our bucket. Solved this by adding today's dat as a folder name and
prepending our filename with that. The next thing we did via
crypto.randombytes is create a random string to save our file as
on AWS. the last thing we did was promisify both s3.upload and
crypto.randomBytes so both implement promises that are resolved
or rejected based on whether or not they work.